Fixed bug with nested [] evaluation.
* Moved builtins to their own files for better test-coverage
  * None added yet.
* Improved handling of "-" when not prefixing a number.
